+++ Monstre.h
@@ -22,7 +22,7 @@ public:
     Monstre(bool degDistance,const QString&  image, pair<double, double> position, double current_hp, double max_hp, double speed, double dmg, QTimer *gameTimer,Player *player=0,QGraphicsScene *scene=0,QGraphicsItem *parent = 0);
     Monstre(bool initNoCreation,bool degDistance,const QString&  image, pair<double, double> position, double current_hp, double max_hp, double speed, double dmg,QTimer *gameTimer,Player *player=0,QGraphicsScene *scene=0,QGraphicsItem *parent = 0);
     virtual ~Monstre(){};
-    void takeDamages(int dmg);
+    void takeDamages(double dmg);
     void testMort();
     bool getDegDistance();
     void attackPlayer();
